Updated RPM Spec
Changed the systemd dependency to require only the
RPM macros on build and systemd proper on install.
Previously full systemd was a build dep and there was no run dep.

Also moved the sed command to change the shebang
on the python script to the `%prep` section, as it
makes more sense there.